\name{print.randomSubspace}
\alias{print.randomSubspace}
\title{Print a randomSubspace Object}
\description{
This function prints \code{randomSubspace} objects.
}
\usage{
\method{print}{randomSubspace}(x, \dots)
}
\arguments{
  \item{x}{
  Fitted model object of class \code{randomSubspace}.
  This is the result of calling the \code{randomSubspace} function.
  }
  \item{\dots}{unused.}
}
\keyword{utilities}
